fix: move logic to specific files
